+++ models/options/OptionSet.py
@@ -72,7 +72,7 @@ class OptionSet:
 
         # Starting setup
         self.starting_level = get_option_default_value("StartingLevel")
-        self.random_starting_level = -1
+        self.random_starting_stats_level = -1
         self.starting_maxhp = get_option_default_value("StartingMaxHP")
         self.starting_maxfp = get_option_default_value("StartingMaxFP")
         self.starting_maxbp = get_option_default_value("StartingMaxBP")
@@ -284,8 +284,8 @@ class OptionSet:
         # Starting setup
         if "StartingMap" in options_dict:
             self.logic_settings.starting_map = options_dict.get("StartingMap")
-        if "RandomStartingLevel" in options_dict:
-            self.random_starting_level = options_dict.get("RandomStartingLevel")
+        if "RandomStartingStatsLevel" in options_dict:
+            self.random_starting_stats_level = options_dict.get("RandomStartingStatsLevel")
         if "StartingMaxHP" in options_dict:
             self.starting_maxhp = options_dict.get("StartingMaxHP")
         if "StartingMaxFP" in options_dict:
@@ -1062,7 +1062,7 @@ class OptionSet:
         # Starting setup
         basic_assert("StartingMap", int)
         basic_assert("StartingLevel", int)
-        basic_assert("RandomStartingLevel", int)
+        basic_assert("RandomStartingStatsLevel", int)
         basic_assert("StartingMaxHP", int)
         basic_assert("StartingMaxFP", int)
         basic_assert("StartingMaxBP", int)
@@ -1764,7 +1764,7 @@ class OptionSet:
         web_settings["StartingMaxFP"] = self.starting_maxfp
         web_settings["StartingMaxBP"] = self.starting_maxbp
         web_settings["StartingLevel"] = self.starting_level
-        web_settings["RandomStartingLevel"] = self.random_starting_level
+        web_settings["RandomStartingStatsLevel"] = self.random_starting_stats_level
         web_settings["StartingStarPower"] = self.starting_starpower
         web_settings["StartingBoots"] = self.logic_settings.starting_boots
         web_settings["StartingHammer"] = self.logic_settings.starting_hammer